pyproject.toml
[project]
name = "antigravtest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "ta>=0.11.0",
    "torch>=2.9.1",
    "yfinance>=0.2.66",
]
